Refactored confusing names in render::canvas
- Rename render::{Cc}anvas -> render::{Ll}ayers
- Rename render::*::Setup -> render::*::Canvas
- Move render::*::setup -> render::*::Canvas::new()
- Removed some obsolete code

+++ client/src/render/scene/node.vert
@@ -7,12 +7,12 @@ uniform mediump vec3 u_sun;
 uniform mediump float u_brightness;
 
 varying mediump vec2 v_tex_pos;
-varying mediump float v_light;
+varying lowp float v_light;
 
 void main() {
     gl_Position = u_proj * vec4(a_pos, 1.0);
     v_tex_pos = a_tex_pos;
-    highp float sun_magnitude = dot(u_sun, a_normal);
+    mediump float sun_magnitude = dot(u_sun, a_normal);
     sun_magnitude = max(0., min(1., sun_magnitude)) * 0.5 + 0.5;
     v_light = sun_magnitude * max(0.5, u_brightness);
 }
